weekly-project-update-generator/functions/requirements.txt
google-cloud-storage==2.10.0
google-cloud-aiplatform==1.38.1
python-pptx==0.6.23
Pillow==10.1.0
functions-framework==3.4.0
google-cloud-logging==3.8.0
python-dotenv==1.0.0
